chore(cli): commands/generate/cell: Separate async handler
