+++ packages/cli/src/commands/generate/helpers.js
@@ -1,88 +1,17 @@
-import path from 'path'
+// This file should only be asynchronously imported by the CLI (typically by
+// being statically imported by a *Handler.js file that is in turn
+// asynchronously imported by the CLI.)
+//
+// Importing this file has side effects that can't be run until after we've set
+// CWD, plus importing this file statically also makes the CLI startup time
+// much slower
 
 import { paramCase } from 'change-case'
-import fs from 'fs-extra'
-import { Listr } from 'listr2'
 import pascalcase from 'pascalcase'
-import terminalLink from 'terminal-link'
 
-import {
-  recordTelemetryAttributes,
-  isTypeScriptProject,
-} from '@redmix/cli-helpers'
-import { getConfig, ensurePosixPath } from '@redmix/project-config'
-import { errorTelemetry } from '@redmix/telemetry'
+import { isTypeScriptProject } from '@redmix/cli-helpers'
 
-import c from '../../lib/colors.js'
-import { generateTemplate, getPaths, writeFilesTask } from '../../lib/index.js'
-import { prepareForRollback } from '../../lib/rollback.js'
 import { pluralize, isPlural, isSingular } from '../../lib/rwPluralize.js'
-
-/**
- * Returns the full path to a custom generator template, if found in the app.
- * Otherwise the default Redwood template.
- */
-export const customOrDefaultTemplatePath = ({
-  side,
-  generator,
-  templatePath,
-}) => {
-  // default template for this generator: ./page/templates/page.tsx.template
-  const defaultPath = path.join(__dirname, generator, 'templates', templatePath)
-
-  // where a custom template *might* exist: /path/to/app/web/generators/page/page.tsx.template
-  const customPath = path.join(
-    getPaths()[side].generators,
-    generator,
-    templatePath,
-  )
-
-  if (fs.existsSync(customPath)) {
-    return customPath
-  } else {
-    return defaultPath
-  }
-}
-
-/**
- * Reduces boilerplate for generating an output path and content to write to disk
- * for a component.
- */
-// TODO: Make this read all the files in a template directory instead of
-// manually passing in each file.
-export const templateForComponentFile = async ({
-  name,
-  suffix = '',
-  extension = '.js',
-  webPathSection,
-  apiPathSection,
-  generator,
-  templatePath,
-  templateVars,
-  componentName,
-  outputPath,
-}) => {
-  const basePath = webPathSection
-    ? getPaths().web[webPathSection]
-    : getPaths().api[apiPathSection]
-  const outputComponentName = componentName || pascalcase(name) + suffix
-  const componentOutputPath =
-    outputPath ||
-    path.join(basePath, outputComponentName, outputComponentName + extension)
-  const fullTemplatePath = customOrDefaultTemplatePath({
-    generator,
-    templatePath,
-    side: webPathSection ? 'web' : 'api',
-  })
-  const content = await generateTemplate(fullTemplatePath, {
-    name,
-    outputPath: ensurePosixPath(
-      `./${path.relative(getPaths().base, componentOutputPath)}`,
-    ),
-    ...templateVars,
-  })
-  return [componentOutputPath, content]
-}
 
 /**
  * Creates a route path, either returning the existing path if passed, or
@@ -103,19 +32,6 @@ export const pathName = (path, name) => {
   return routePath
 }
 
-const appendPositionalsToCmd = (commandString, positionalsObj) => {
-  // Add positionals like `page <name>` + ` [path]` if specified
-  if (Object.keys(positionalsObj).length > 0) {
-    const positionalNames = Object.keys(positionalsObj)
-      .map((positionalName) => `[${positionalName}]`)
-      .join(' ')
-    // Note space after command is important
-    return `${commandString} ${positionalNames}`
-  } else {
-    return commandString
-  }
-}
-
 /** @type {(name: string, generatorName: string) => string } **/
 export function removeGeneratorName(name, generatorName) {
   // page -> Page
@@ -127,137 +43,11 @@ export function removeGeneratorName(name, generatorName) {
   return coercedName
 }
 
-/** @type {Record<string, import('yargs').Options>} */
-export const yargsDefaults = {
-  force: {
-    alias: 'f',
-    default: false,
-    description: 'Overwrite existing files',
-    type: 'boolean',
-  },
-  typescript: {
-    alias: 'ts',
-    default: isTypeScriptProject(),
-    description: 'Generate TypeScript files',
-    type: 'boolean',
-  },
-}
-
 export const validateName = (name) => {
   if (name.match(/^\W/)) {
     throw new Error(
       'The <name> argument must start with a letter, number or underscore.',
     )
-  }
-}
-
-/**
- * Reduces boilerplate for creating a yargs handler that writes a
- * component/page/layout/etc to a location.
- */
-export const createYargsForComponentGeneration = ({
-  componentName,
-  preTasksFn = (options) => options,
-  /** filesFn is not used if generator implements its own `handler` */
-  filesFn = () => ({}),
-  optionsObj = yargsDefaults,
-  positionalsObj = {},
-  /** function that takes the options object and returns an array of listr tasks */
-  includeAdditionalTasks = () => [],
-}) => {
-  return {
-    command: appendPositionalsToCmd(`${componentName} <name>`, positionalsObj),
-    description: `Generate a ${componentName} component`,
-    builder: (yargs) => {
-      yargs
-        .positional('name', {
-          description: `Name of the ${componentName}`,
-          type: 'string',
-        })
-        .epilogue(
-          `Also see the ${terminalLink(
-            'Redwood CLI Reference',
-            `https://redwoodjs.com/docs/cli-commands#generate-${componentName}`,
-          )}`,
-        )
-        .option('tests', {
-          description: 'Generate test files',
-          type: 'boolean',
-        })
-        .option('stories', {
-          description: 'Generate storybook files',
-          type: 'boolean',
-        })
-        .option('verbose', {
-          description: 'Print all logs',
-          type: 'boolean',
-          default: false,
-        })
-        .option('rollback', {
-          description: 'Revert all generator actions if an error occurs',
-          type: 'boolean',
-          default: true,
-        })
-
-      // Add in passed in positionals
-      Object.entries(positionalsObj).forEach(([option, config]) => {
-        yargs.positional(option, config)
-      })
-      // Add in passed in options
-      Object.entries(optionsObj).forEach(([option, config]) => {
-        yargs.option(option, config)
-      })
-    },
-    handler: async (options) => {
-      recordTelemetryAttributes({
-        command: `generate ${componentName}`,
-        tests: options.tests,
-        stories: options.stories,
-        verbose: options.verbose,
-        rollback: options.rollback,
-        force: options.force,
-        // TODO: This does not cover the specific options that each generator might pass in
-      })
-
-      if (options.tests === undefined) {
-        options.tests = getConfig().generate.tests
-      }
-      if (options.stories === undefined) {
-        options.stories = getConfig().generate.stories
-      }
-      validateName(options.name)
-
-      try {
-        options = await preTasksFn(options)
-
-        const tasks = new Listr(
-          [
-            {
-              title: `Generating ${componentName} files...`,
-              task: async () => {
-                const f = await filesFn(options)
-                return writeFilesTask(f, { overwriteExisting: options.force })
-              },
-            },
-            ...includeAdditionalTasks(options),
-          ],
-          {
-            rendererOptions: { collapseSubtasks: false },
-            exitOnError: true,
-            renderer: options.verbose && 'verbose',
-          },
-        )
-
-        if (options.rollback && !options.force) {
-          prepareForRollback(tasks)
-        }
-        await tasks.run()
-      } catch (e) {
-        errorTelemetry(process.argv, e.message)
-        console.error(c.error(e.message))
-        process.exit(e?.exitCode || 1)
-      }
-    },
   }
 }
