+++ packages/cli/src/commands/generate/model/model.js
@@ -13,7 +13,8 @@ import {
 } from '../../../lib/index.js'
 import { prepareForRollback } from '../../../lib/rollback.js'
 import { verifyModelName } from '../../../lib/schemaHelpers.js'
-import { validateName, yargsDefaults } from '../helpers.js'
+import { validateName } from '../helpers.js'
+import { yargsDefaults } from '../yargsCommandHelpers.js'
 const TEMPLATE_PATH = path.resolve(__dirname, 'templates', 'model.js.template')
 
 const files = async ({ name, typescript = false }) => {
