+++ packages/cli/src/commands/generate/scaffold/__tests__/scaffold.test.js
@@ -9,7 +9,7 @@ import { vi, describe, test, expect, afterAll, beforeAll } from 'vitest'
 import '../../../../lib/test'
 
 import { getDefaultArgs } from '../../../../lib/index.js'
-import { yargsDefaults as defaults } from '../../helpers.js'
+import { yargsDefaults as defaults } from '../../yargsCommandHelpers.js'
 import * as scaffold from '../scaffold.js'
 
 vi.mock('fs-extra', async (importOriginal) => {
